// src/shared/api/models/patchApiServicesDepositBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PatchApiServicesDepositBodyDepositSetting } from './patchApiServicesDepositBodyDepositSetting';
import type { PatchApiServicesDepositBodyServicesItem } from './patchApiServicesDepositBodyServicesItem';

export type PatchApiServicesDepositBody = {
    depositSetting: PatchApiServicesDepositBodyDepositSetting;
    /** @minItems 1 */
    services: PatchApiServicesDepositBodyServicesItem[];
};